insns.dat: Minor corrections after XED validation
Issues found after comparing instruction encodings with the intel-xed
tool.

MOVSXD's second operands is 32-bit only when the first operand is either
32-bit or 64-bit. If the first operand is 16-bit then so should the
second one - based on the SDM.

ARPL didn't assemble for "ARPL rm16, reg32" due to the SM flag. Remove
it.

SKINIT uses the EAX register but NASM doesn't require it as an operand
when assembling. Add a version with the operand so the intel-xed
disassembled SKINIT can be assembled with NASM.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>