tools: Script to validate all instructions agains intel-xed
Write a python script that will check all the encodings agains a second
tool such as intel-xed.

The check requires first assembling an instruction with NASM, then
disassembling it with XED and taking that disassembled piece of assembly
and assembling it again with NASM to verify the encodings are equal.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>

testing

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>